Estimate causal tail coefficient

Estimates the causal tail coefficient between two variables v1 and v2, given the threshold k.

causal_tail_coeff(v1, v2, k = floor(n^0.4), to_rank = TRUE, both_tails = TRUE)

Arguments

v1, v2

Numeric vectors. Two vectors with n observations.

k

Positive integer. The number of extreme observations used to compute the causal tail coefficient. Set by default to k = floor(n^0.4). It must be greater than 1 and smaller than n.

to_rank

Boolean. Are the vectors v1 and v2 already sorted? Set by default to TRUE.

both_tails

Boolean. Do you consider both tails when computing the causal tail coefficient? If both_tails = TRUE, then you use the \(\Psi\) formulation. If both_tails = FALSE, then you use the \(\Gamma\) formulation. Set by default to TRUE.

Value

Numeric --- between 0 and 1. The causal tail coefficient between v1 and v2.

Details

The causal tail coefficient is defined in the paper "Causal discovery in heavy-tailed models" and has two formulations.

The first formulation is defined in the paper as the \(\Gamma\)-coefficient, and it considers only the upper tails of the variables. To use this formulation set both_tails = FALSE.

The second formulation is defined in the paper as the \(\Psi\)-coefficient, and considers both tails of the variables. To use this formulation set both_tails = FALSE.

In general, the causal tail coefficient is not symmetric, i.e., causal_tail_coeff(v1, v2) != causal_tail_coeff(v2, v1).