Simulate noise observations

Sample n observations for p independent noise variables from a certain distribution distr.

simulate_noise(
  n,
  p,
  distr = c("student_t", "gaussian", "log_normal"),
  tail_index
)

Arguments

n

Positive integer. The number of observations, must be larger than 1.

p

Positive integer. The number of variables, must be larger than 1.

distr

Character. The distribution of the noise. It is one of:

student_t, in this case the user has to specify the tail_index, i.e., the degrees of freedom,

gaussian,

log_normal.

tail_index

Positive numeric. The tail index, i.e., degrees of freedom, of the noise.

Value

Numeric matrix. Dataset matrix with n

rows (observations) and p columns (variables).